requirements.txt
et_xmlfile==2.0.0
lxml==6.0.2
openpyxl==3.1.5
pillow==12.1.1
python-docx==1.2.0
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
